refac: Extract CLI parser into cli.rs
